refactor(Price): move domain to menu package
