+++ src/main/java/kitchenpos/application/MenuService.java
@@ -10,7 +10,7 @@ import kitchenpos.dao.MenuProductDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.Price;
+import kitchenpos.domain.menu.Price;
 import kitchenpos.domain.menu.Product;
 import kitchenpos.domain.menu.ProductQuantity;
 import org.springframework.stereotype.Service;
