+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -1,7 +1,6 @@
 package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Price;
 
 public class Menu {
