+++ src/main/java/kitchenpos/domain/menu/Price.java
@@ -1,8 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.menu.ProductQuantity;
 
 public class Price {
